Implement libraries of threads for each sensor on GY-80
The four sensors: accelerometer, gyroscope, magnetometer and
barometric pressure can be used as library. These libraries
are used, each one, in one thread. There is a fiveth
thread in which is get all the results of each sensor. The five
threads are managed by a threadController that allows to run all
threads at once.